Fix mirrorcult suggestions
