+++ Content.Shared/Storage/Components/ItemCounterComponent.cs
@@ -1,4 +1,6 @@
-﻿using Content.Shared.Whitelist;
+﻿using Content.Shared.Storage.EntitySystems;
+using Content.Shared.Whitelist;
+using Robust.Shared.Analyzers;
 using Robust.Shared.GameObjects;
 using Robust.Shared.Serialization.Manager.Attributes;
 
@@ -17,6 +19,7 @@ namespace Content.Shared.Storage.Components
     ///      tags: [Cigarette]
     /// </code>
     [RegisterComponent]
+    [Friend(typeof(SharedItemCounterSystem))]
     public class ItemCounterComponent : Component
     {
         public override string Name => "ItemCounter";
