+++ Content.Shared/Storage/Components/ItemMapperComponent.cs
@@ -1,4 +1,6 @@
 ﻿using System.Collections.Generic;
+using Content.Shared.Storage.EntitySystems;
+using Robust.Shared.Analyzers;
 using Robust.Shared.GameObjects;
 using Robust.Shared.Serialization;
 using Robust.Shared.Serialization.Manager.Attributes;
@@ -6,6 +8,7 @@ using Robust.Shared.Serialization.Manager.Attributes;
 namespace Content.Shared.Storage.Components
 {
     [RegisterComponent]
+    [Friend(typeof(SharedItemMapperSystem))]
     public class ItemMapperComponent : Component, ISerializationHooks
     {
         public override string Name => "ItemMapper";
